+++ packages/core/release.config.mjs
@@ -12,36 +12,23 @@ export default {
             {
                 preset: "conventionalcommits",
                 presetConfig: {
-                    headerPattern: /^(\w*)(?:\((.*)\))?!?: (.*)$/,
-                    headerCorrespondence: ["type", "scope", "subject"],
-                    noteKeywords: ["BREAKING CHANGE"],
                     types: [
                         { type: "feat", section: "Features" },
                         { type: "fix", section: "Bug Fixes" },
                         { type: "chore", section: "Chores" },
                         { type: "docs", section: "Documentation" },
-                        { type: "style", section: "Styles" },
-                        { type: "refactor", section: "Code Refactoring" },
-                        { type: "perf", section: "Performance Improvements" },
-                        { type: "test", section: "Tests" }
-                    ]
+                    ],
                 },
                 writerOpts: {
-                    headerPartial: `## {{version}} - {{date}}\n`,
-                    transform: (commit, context) => {
-                        commit.version = context.version;
-                        commit.date = context.date;
-                        return commit;
-                    },
-                }
-            }
+                    headerPartial: `## hello {{version}} - {{date}}`, // This controls the version title format in the changelog
+                },
+            },
         ],
-        "@semantic-release/release-notes-generator",
         [
             "@semantic-release/changelog",
             {
                 changelogFile: "./CHANGELOG.md",
-                changelogTitle: "# Changelog for @geniux/test-semantic-release-core",
+                changelogTitle: "# Changelog for\n\n @geniux/test-semantic-release-core",
             },
         ],
         // "@semantic-release/npm",